requirements.txt
# Deep Learning Frameworks
torch==2.2.0
torchvision==0.17.0
transformers==4.36.0

# Parameter-Efficient Fine-Tuning
peft==0.7.1
accelerate==0.24.0
bitsandbytes==0.41.3

# Data Processing
datasets==2.14.0
pandas==2.2.0
numpy==1.26.0
scikit-learn==1.3.0

# Financial Data APIs
sec-api==1.0.17
yfinance==0.2.32

# NLP Tools
nltk==3.8.1
spacy==3.7.2
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl

# Visualization
matplotlib==3.8.0
seaborn==0.12.2

# Experiment Tracking
wandb==0.16.0

# Web Application
gradio==4.7.1

# Utilities
tqdm==4.66.1
python-dotenv==1.0.0
